Move c++ type parser to separate file
